Fix export-import issue
