+++ src/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Bulls & Cows</title>
     <link rel="stylesheet" href="./styles/main.css">
+    <script type="module" src="scripts/main.js"></script>
   </head>
   <body>
     <h1>Bulls & Cows</h1>
@@ -60,6 +61,6 @@
         <p class="input__error"></p>
       </section>
     </div>
-    <script type="text/javascript" src="scripts/main.js"></script>
+
   </body>
 </html>
